OPC server page is available only for admin
