+++ src/main/java/com/nppgks/reportingsystem/controller/OpcServiceController.java
@@ -1,10 +1,8 @@
 package com.nppgks.reportingsystem.controller;
 
 import com.nppgks.reportingsystem.opcservice.OpcServiceRequests;
-import com.nppgks.reportingsystem.service.dbservices.AllTagNamesService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
@@ -13,8 +11,6 @@ import org.springframework.web.bind.annotation.*;
 public class OpcServiceController {
 
     private final OpcServiceRequests opcServiceRequests;
-    private final AllTagNamesService allTagNamesService;
-
     @GetMapping("/testOpcServerConnection")
     @ResponseBody
     public boolean testConnection(){
@@ -31,11 +27,5 @@ public class OpcServiceController {
     @ResponseBody
     public String getValueByName(@PathVariable String tagName){
         return opcServiceRequests.getTagDataFromOpc(tagName);
-    }
-
-    @GetMapping("/opcServer")
-    public String getOpcServerPage(ModelMap modelMap){
-        modelMap.put("tagNames", allTagNamesService.getAllOperativeAndCalculationTagNames());
-        return "opc-server";
     }
 }
